+++ blobs.js
@@ -3,14 +3,16 @@ let blobAmount = 5;
 
 function setup() {
 	let canvas = createCanvas(
-		window.innerWidth /2,
-		window.innerHeight /2
+		window.innerWidth /1.5,
+		window.innerHeight /1.5
 		);
 
 	canvas.parent('sketch-holder');
 
+	colorMode(HSB);
+
 	for (let i=0; i<blobAmount; i++) {
-		blobs[i] = new Blob(10, color(23, 140, 240));
+		blobs[i] = new Blob(randomGaussian(20, 2), color(random(255), 50, 90));
 	}
 }
 
@@ -19,5 +21,27 @@ function draw() {
 	blobs.forEach(function(blob) {
 		blob.run();
 	});
-	
+	checkForCollisions();
+}
+
+function checkForCollisions() {
+	blobs.forEach(function(blob) {
+		blobs.forEach(function(otherBlob) {
+			if (otherBlob == this) {
+				return;
+			}
+			if(blob.isCollidingWith(otherBlob)) {
+				if (blob.size > otherBlob.size) {
+					blob.eat(otherBlob);
+					otherBlob.die();
+				} else if (blob.size < otherBlob.size) {
+					otherBlob.eat(blob);
+					blob.die();
+				} else {
+					blob.bounceFrom(otherBlob);
+					otherBlob.bounceFrom(Blob);
+				}
+			}
+		});
+	});
 }