+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- Blobs roam with random velocity and color, the larger eating the smaller on collision.
+- Centering sketch in window with css.
+- Eaten blobs are declared dead and are no longer updated or displayed.
 
 ## [0.1.0] - 2016-12-18
 ### Added